Refactor indexes.
