+++ src/test/java/com/cosyan/db/index/ByteTrieTest.java
@@ -23,8 +23,8 @@ import java.nio.file.Paths;
 import org.junit.Test;
 
 import com.cosyan.db.index.ByteTrie.IndexException;
-import com.cosyan.db.index.ByteTrie.LongIndex;
-import com.cosyan.db.index.ByteTrie.StringIndex;
+import com.cosyan.db.index.LongLeafTries.LongIndex;
+import com.cosyan.db.index.LongLeafTries.StringIndex;
 
 public class ByteTrieTest {
